Move Mlir lowering to targets
This commits moves the lowering passes for LLVMIR and C targets from
common Mlir files to target specific files.
The new _lowering_pipeline methods will allow dynamic pipelines in the
future.
