+++ src/xtc/backends/mlir/MlirTarget/MlirLLVMTarget.py
@@ -32,7 +32,8 @@ from xtc.itf.graph import Graph
 from .MlirTarget import MlirTarget
 from ..MlirConfig import MlirConfig
 from ..MlirProgram import RawMlirProgram
-from ..MlirCompilerPasses import MlirProgramToLLVMDialectPass
+
+from mlir.passmanager import PassManager
 
 __all__ = ["MlirLLVMTarget"]
 
@@ -251,3 +252,58 @@ class MlirLLVMTarget(MlirTarget):
         else:
             result = subprocess.run(cmd, text=True)
         return result
+
+
+class MlirProgramToLLVMDialectPass:
+    def __init__(
+        self,
+        mlir_program: RawMlirProgram,
+    ) -> None:
+        self._mlir_program = mlir_program
+
+    def _lowering_pipeline(self) -> list[str]:
+        return [
+            "canonicalize",
+            "cse",
+            "sccp",
+            # From complex control to the soup of basic blocks
+            "expand-strided-metadata",
+            "convert-linalg-to-loops",
+            "lower-affine",
+            "convert-vector-to-scf{full-unroll=true}",
+            "scf-forall-to-parallel",
+            "convert-scf-to-openmp",
+            "canonicalize",
+            "cse",
+            "sccp",
+            "convert-scf-to-cf",
+            "canonicalize",
+            "cse",
+            "sccp",
+            # Memory accesses to LLVM
+            "buffer-results-to-out-params",
+            "convert-func-to-llvm{use-bare-ptr-memref-call-conv=true}",
+            "finalize-memref-to-llvm",
+            "canonicalize",
+            "cse",
+            "sccp",
+            # Data flow to LLVM
+            "convert-vector-to-llvm{enable-x86vector=true}",
+            "convert-index-to-llvm",
+            "convert-arith-to-llvm",
+            "canonicalize",
+            "cse",
+            "sccp",
+            # Control flow to LLVM
+            "convert-cf-to-llvm",
+            "convert-openmp-to-llvm",
+            "canonicalize",
+            "cse",
+            "sccp",
+        ]
+
+    def run(self) -> None:
+        pm = PassManager(context=self._mlir_program.mlir_context)
+        for opt in self._lowering_pipeline():
+            pm.add(opt)  # type: ignore # no attribte add?
+        pm.run(self._mlir_program.mlir_module.operation)
