+++ src/xtc/utils/ext_tools.py
@@ -37,71 +37,6 @@ transform_opts = [
     "transform-interpreter",
 ]
 
-lowering_llvm_opts = [
-    "canonicalize",
-    "cse",
-    "sccp",
-    # From complex control to the soup of basic blocks
-    "expand-strided-metadata",
-    "convert-linalg-to-loops",
-    "lower-affine",
-    "convert-vector-to-scf{full-unroll=true}",
-    "scf-forall-to-parallel",
-    "convert-scf-to-openmp",
-    "canonicalize",
-    "cse",
-    "sccp",
-    "convert-scf-to-cf",
-    "canonicalize",
-    "cse",
-    "sccp",
-    # Memory accesses to LLVM
-    "buffer-results-to-out-params",
-    "convert-func-to-llvm{use-bare-ptr-memref-call-conv=true}",
-    "finalize-memref-to-llvm",
-    "canonicalize",
-    "cse",
-    "sccp",
-    # Data flow to LLVM
-    "convert-vector-to-llvm{enable-x86vector=true}",
-    "convert-index-to-llvm",
-    "convert-arith-to-llvm",
-    "canonicalize",
-    "cse",
-    "sccp",
-    # Control flow to LLVM
-    "convert-cf-to-llvm",
-    "convert-openmp-to-llvm",
-    "canonicalize",
-    "cse",
-    "sccp",
-]
-
-lowering_std_opts = [
-    "canonicalize",
-    "cse",
-    "sccp",
-    # From complex control to the soup of basic blocks
-    "expand-strided-metadata",
-    "convert-linalg-to-loops",
-    "lower-affine",
-    "convert-vector-to-scf{full-unroll=true}",
-    "scf-forall-to-parallel",
-    "convert-scf-to-openmp",
-    "canonicalize",
-    "cse",
-    "sccp",
-    # "convert-scf-to-cf",
-    "canonicalize",
-    "cse",
-    "sccp",
-    # Memory accesses
-    "buffer-results-to-out-params",
-    "canonicalize",
-    "cse",
-    "sccp",
-]
-
 mlirtranslate_opts = ["--mlir-to-llvmir"]
 
 xtctranslate_opts = ["--mlir-to-c"]
